Used shared call to bind resources to an EP
Create a shared version of bind_ep_res(), which is duplicated
by most apps to some degree.  There are a couple of instances
where a shared call cannot easily be used, such as the scalable
endpoint test.  Those are left unchanged.

About half the apps post an initial receive buffer in
bind_ep_res().  Use a parameter to indicate if a receive
should be posted.  Subsequent work will examine if all apps
can be made consistent in this area.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>