Move active EP allocation to common code
Replace individual implementations of alloc_ep_res() with
a shared implementation for allocating an active EP and
associated resources.

This patch also fixes incorrect usage of determining the
'max_credits' of requests that can be posted to the transmit
queue of the EP.

Private usages of 'transfer_size' are also replaced with
the common transfer size option.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>