Move init_av calls into common code
Replace the dozen or so init_av() duplicate definitions of init_av()
with a common implementation.

In order to insert the remote address of the client side of the
test, a message exchange is needed to provide the server with the
client address.  This requires moving the initial control message
exchange into common code.

That address exchange trickles throughout the code in order to
ensure that receive buffers are posted before sends are initiated,
provide enough buffer space for addressing, and handle FI_PREFIX
correct (for those apps that already support it).

Signed-off-by: Sean Hefty <sean.hefty@intel.com>