Cleanup atomic definitions
The PSM and sockets provider differ in their implementations
of the atomic compare-swap operations.  For example, the PSM
provider does this for CSWAP_LT:

   if remote buffer < compare value then
      remote buffer = data value

whereas the sockets provider implements:

   if compare value < remote buffer then
      remote buffer = data value

Other CSWAP functions are similar.  The PSM provider matches
the implementation defined by the man pages.  However, this
gets more confusing because the man pages define this for MIN

   if compare value < remote buffer then
      remote buffer = data value

MAX is similar.  So the man pages are inconsistent in how atomics
are defined, and the providers differ in the implementation.  The
'natural' or expected behavior for atomics is more closely aligned
with the MIN/MAX definition than the CSWAP definitions.

We use the socket provider implementation as the 'correct' behavior.
The man pages are updated to reflect the actual behavior, with the
PSM provider updated to support the new definitions.

There are no known users of the atomic definitions at this point,
so this should be a relatively 'safe' change.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>