Cleanup and fix rdm_multi_recv completion handling
fi_rdm_multi_recv assumes that FI_MULTI_RECV completions
are always separate from data completions.  This is not
always the case, for example, the PSM provider reports
this as part of the last data completion.  Update the
app to test for this correctly.

Provide a broad cleanup of the code.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>