Move fi_domain() into ft_open_fabric_res()
Nearly all tests open the domain immediately after opening
the fabric resources (fabric and eq).  Move opening the
domain into the common code.

There are a couple of exceptions to this.  The server
side of connected endpoints defer opening the domain until
the connreq is received.  And a couple of unit tests open
the domain differently.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>