Use common code for message processing
Provide a common implementation for sending, receiving,
and completion process.
This will help all test apps support (with additional changes)
the FI_PREFIX and FI_CONTEXT mode bits.

As part of this change, we assign a virtual tx sequence number
to each sent message.  The receiving side matches this with
an rx sequence number.  The sequence numbers are used as tags
when dealing with the tag matching interface.

Completion processing is updated to count the number of entries
that have been retrieved from a CQ.  Calls are updated to wait
until a total number of completions have occurred, rather than
waiting for X new completions.  This will allow CQs to be
converted to counters more easily.

This change fixes several places where fi_context structures
were reused before a previous request completed.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>