Add common call to post a send message
Provide a common implementation for sending a
message and waiting for a completion.  This will
help all test apps support (with additional changes)
the FI_PREFIX and FI_CONTEXT mode bits.

As part of this change, track the number of transmit
requests that have been initiated and the number which
have completed.  This will allow us to use counters
to wait for operations to complete.

The common send implementation uses a send then wait
for completion mechanism.  This is necessary to ensure
that we don't overwrite either the send buffer or
transmit context (fi_context) while the send is in
operation.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>