man: Remove reference to RMA generating a CQ entry
The man pages are unclear on FI_RMA_EVENT generating a
CQ entry in response to an RMA operation.  The issue is
that only counters were defined to respond to RMA events
(see fi_ep_bind).  Support for CQ events was removed from
the 1.0 release because of the negative impact to the
providers, but the man pages were left in a confusing
state.

Addresses issue #1300.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>