prov/psm: reduce the default timeout value for psm_ep_close
The first call to psm_ep_close is with the graceful flag and tries
to do the cleanup job nicely. When it fails (timeout expired) it
is called again with the forced flag. The old default timeout value
(120 seconds) is too long and can cause large scale application to
hang for hours if something goes wrong in the finalization phase.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>