prov/psm: add a lock to the polling function for thread safety
Protect the sequence "psm_mq_ipeek; psm_mq_test" with a lock so
that they cannot be interleaved when issued from multiple threads.
Without the protection, the call to psm_mq_test could try to
access an request that has already been freed.

This enables the addition of an auto progress thread.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>