common: Delay setting default port values
The src and dst port values are initialized to the same value
during test startup.  This results in failures when running on
the same system and binding to a specific address.  For example:

fi_cq_data -s 127.0.0.1
fi_cq_data -s 127.0.0.1 127.0.0.1

fails on the server calling fi_accept().  The failure traces
back to a failed socket bind() call.

Fix this by delaying when the default port values are set.
Set the src and/or dst port in ft_read_addr_opts(), when
processing the test options.  Update all tests to use this
function for consistency.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>