prov/psm: add option to set the CPU affinity of the progress thread
For applications that purely rely on auto progress, the performance
can be quite bad if the main thread and the progress thread are
aggresively competing with each other (e.g. the main thread keeps
polling message buffer to detect completion). Setting CPU affinity
for the progress thread can lead to significant improvement.

The new environment variable "FI_PSM_PROG_AFFINITY" (or "FI_PSM2_
PROG_AFFINITY" for the psm2 provider) controls the affinity setting
of the progress thread. If the value is 0, affinity is disabled.
If the value is positive, the thread is pinned to the core counting
from the lowest numbered core (i.e. "1" means core 0). If the value
is negative, the thread is pinned to the core counting from the
highest numbered core. The default is "-1", which means the last
core.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>