prov/psm: new format for progress thread affinity setting
Replace the single number used for the affinity setting (for
variable FI_PSM_PROG_AFFINITY or FI_PSM2_PROG_AFFINITY) with
a sequence of triplets <start>:<end>:<stride> separated by
commas. The value <start> and <end> can be either the actual
"core_id" (when >=0) or "core_id - num_cores" (when <0). The
latter is a convenient way to specify "the i-th core from the
last".

The affinity is not set if the variable is undefined.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>