usnic: fix mode handling
Based on the clarification in e937c58eec3f40566a5e122ab58355dd8aa1b204,
we should not be interpreting (hints != NULL && hints->[...]->mode == 0)
as a request to set the modes desired by the provider.  Instead we
should assume that an explicit "hints" structure with zero mode values
is indicating no modes are supported by the application.

Signed-off-by: Dave Goodell <dgoodell@cisco.com>