prov/psm: fix the handling of context for ops w/o completion
According to the man pages, on EPs with FI_SELECTIVE_COMPLETION
flag set, the "context" parameter should be ignored for operations
without the FI_COMPLETION flag.

Previously, for recv ops only NULL context was ignored in such
case. The intention was to allow non-NULL context be passed for
the purpose of calling fi_cancel later. However, such behavior
doesn't conform to the spec and has the possibility of causing
memory corruption.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>