man/cq: Fix description of EPs bound to CQs and counters
The man pages states that a EP that is bound to both a CQ
and a counter will not generate completions on the CQ except
in error cases.  The text should reference the use of the
FI_SELECTIVE_COMPLETION flag for supressing completions.
An EP bound to a CQ *will* generate completions by default
unless completions are disabled.

This updates the man page to match the desired intent, as
well as the implementation provided by the sockets and psm
providers.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>