usnic: improve fi_cq_sread performance
The usnic provider's fi_cq_sread performance was suboptimal before this
change.  It used a static 5 millisecond (not microsecond) sleep time
between polling the underlying CQ, resulting in terrible performance for
use cases where a completion would be arriving soon but was not actually
present yet in the CQ to be read upon initial entry into fi_cq_sread.

This commit changes the implementation to use an exponential back-off
strategy that starts small (2 microseconds) but doubles up to a cap
(currently 5 milliseconds).

Signed-off-by: Dave Goodell <dgoodell@cisco.com>