prov/verbs: Make a copy for check_info
After finding the right "verbs_info", make a copy of it for
check_info so that check_info->next is NULL. This is needed
so that fi_ibv_get_matching_info doesn't return a list of
fi_info structures. When the info given in hints resolves to
a particular verbs device the app expects only one fi_info
structure.

Signed-off-by: Arun C Ilango <arun.ilango@intel.com>