ubertest: Simplify EAGAIN inject testing loop
The inject testing cannot wait for a completion, so relies on
checking for EAGAIN.  The test loop currently calls 'transmit',
followed by checking for completions (to force progress, if
needed), and calls fi_tx_size_left.  The size_left call is
unneeded and can add additional latency.  Remove it and just
loop on the transmit calls.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>