eliminate temporaries in atomic macros
is there a reason to load the dest data into a temp and only
store it into the result after the atomic operation has
been evaluated?  the atomic op should not fail unless the
user has provided bad buffer (i.e. it will segfault)
and the elemental operations are not atomic.

perhaps it is to impede compiler optimizations?