prov/gni: Replace mbox double_free tests with one
that checks for errors.

I think the intention of the double_free test was to check an error
path, but as written, it reads freed memory and cannot be relied on
(though it usually works).  Replace this with a new test that forces
_gnix_mbox_free() to exercise the error paths by manipulating internal
state.

Upstream merge of ofi-cray/libfabric-cray#505

Signed-off-by: Sung-Eun Choi <sungeunchoi@users.noreply.github.com>
(cherry picked from commit ofi-cray/libfabric-cray@cb4e58da5843f7c3289a4839b0100973294348b0)