prov/psm: support running psm provider over psm2-compat library
The AM handler has different signature in PSM and PSM2. That
is the reason why the psm provider can't run with the psm2-compat
library even if the library is ABI compatible with the PSM
library (AM is not a public interface in PSM).

This patch adds an AM compatibility layer to the psm provider
so that it can work with both the original PSM library and the
psm2-compat library (on the respective supported HW, of course).

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>