prov/psm: add parameter for progress thread suppression
By default, the progress thread is created if the application asks
for auto progress support. The progress thread has minor but still
measurable impact on the performance and the effect can be more
noticeable on certain platforms with slower cores. There are cases
that auto progress is not really needed but the application still
asks for it blindly. In such cases, turning of the progress thread
can improve the performance without affecting the functionality.

The parameter FI_PSM_PROG_THREAD / FI_PSM2_PROG_THREAD determines
if the progress thread is allowed to be created. When set to "no",
no progress thread is created even if auto progress is asked for.
The default setting is "yes" which means the progress thread can
be created when needed.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>