configure: list all valid settings for the "--enable-xxx" options
The "--enable-xxx" options (where "xxx" is the provider name) have
more choices than the standard yes|no|auto values. Listing all the
possible choices in the help text can be useful to new users.

As a special note, the option format "--enable-xxx=<PATH>" is the
equivalence of the "--with-xxx=<PATH>" option found elsewhere.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>