travis: install dependencies for the verbs provider
Inspired by the work previously done in #1360.

Explicitly request that the verbs and sockets providers be built, since
it should be a bug if they fail to configure under our new Travis
settings.

This theoretically should enable the usnic provider to build too, but
there appears to be a bug in the libnl-3 configury logic around the
`nl_recvmsgs_report` symbol not being present in the version available
for the Travis Ubuntu environment (libnl 3.2.3).

Fixes #1538

Signed-off-by: Dave Goodell <dgoodell@cisco.com>