Add completion method that polls on the fd wait object
Some applications (e.g. rsockets, ES-API) rely on the ability
to call poll/select on the underlying fd associated with a
completion queue.  Add a new completion mechanism to test
whether a provider supports this correctly.  This also shows
how fd processing should be done to ensure that the signal
associated with the fd is properly reset after being signaled.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>