usnic: one last condense down to $(usnic_LIBS)
The usnic provider needs to -libverbs if the usnic or verbs providers
are built as DLs.

The determination of whether usnic needs to -libverbs used to be made
as an AM_CONDITIONAL in Makefile.am (i.e., after all the providers had
been configured, and we could easily check whether verbs or usnic were
being built as DLs).  However, to properly support pkg-config, we need
to make this -libverbs-or-not decision when the usnic provider is
configured.

The only real way to do that is just to ensure that the verbs provider
is configured *before* the usnic provider.  This commit therefore adds
a comment in configure.ac explaining this (which is hackish, but there
isn't really a better solution).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>