build system: move provider-specific Makefilery to their dirs
Continue the trend started by the gnu provider: move all the
provider-specific Makefile.am bits into
prov/PROVIDER_NAME/Makefile.include, and then just include those files
in the top-level Makefile.am.

Also rename the gni Makefile.am to Makefile.include (like the rest of
the included Makefiles) so that it's clear from the filename that this
is not a normal Makefile.am file -- it must be included to be used.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>