mxm: condense down to $(mxm_CPPFLAGS|LDFLAGS|LIBS)
Update the mxm provider configury to set mxm_CPPFLAGS, mxm_LDFLAGS,
and mxm_LIBS, and then use those flags in the mxm provider's
Makefile.include.

This change is necessary for the upcoming pkg-config support.

Also, update the MXM Makefile.include to properly set flags and libs
in the non-DL case.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>