prov/psm2: remove AM based message implementation
The main purpose of the AM based message fucntions is to provide
an alternative implementation that doesn't take away usable tag
bits from tagged message functions. This is no longer a concern
in PSM2.

Furthermore, the AM based message functions is not as full featured
as the regular one, and the performance is not as good. Removing
this part not only reduces the code size, but also eliminates a
conditional branch in the progress function.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>