usnic: gracefully handle when nl_send() temporarily fails
Sometimes nl_send() can fail simply because the kernel is temporarily
out of resources, and we should just try again.  libnl1 and libnl3
handle this case a little differently, so use the USD_NL_SEND() macro
to hide the differences.  If retry comes back as true, then sleep a
little and try again.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>