usnic: allow FI_WAIT_UNSPEC
Also, deliberately error (FI_EOPNOTSUPP) on calls to fi_cq_sread when
FI_WAIT_NONE has been specified.  It's not as user friendly in some
ways, but I think taking a stricter interpretation of the libfabric docs
is helpful to users and the project in the long run so that users do not
get the incorrect impression that it is OK to use NONE+sread together.

See discussion in #1629 for more info.

Signed-off-by: Dave Goodell <dgoodell@cisco.com>