UDP: Add UDP provider with common implementation
This introduces a UDP sockets provider, along with a utility 'library'.
The UDP provider is a DGRAM based provider that sends and receives
UDP packets. It should be compatible with the usnic provider, but I
haven't tested that. The UDP provider is built using basic provider
building blocks -- the utility 'library'. The reasoning for
introducing a UDP provider is that it provides the lowest level of
functionality over which other features can be constructed and
added to the utility library.

The utility library is intended to serve two purposes. The first is
to provide common building blocks that all providers can use.
This will allow us to focus efforts on optimizing a single implementation
of various features that any provider can to use. For example, the
utility code provides basic definitions for libfabric objects in order
to handle reference counting, check attributes, etc. It also provides a
software implementation of CQs, EQs, pollsets, AVs, ... I anticipate
developers iterating over the utility library code as providers are
updated to use it.

The utility library is based on existing code from the sockets and
psm providers. Performance is a primary focus, and it expects provider
specific code segments as a result. For example, the EP and CQ code are
almost entirely provider specific.

The intent is to extend the utility library to implement additional
libfabric features (such as tagged interfaces, multi-recv support,
etc.), plus construct a utility provider that can automatically
upgrade the feature set of an existing provider.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>