udp: Switch to support only for sockaddr_in
Replace FI_SOCKADDR with FI_SOCKADDR_IN in fi_getinfo output.
Address vectors need to know the specific address format
they are dealing with to optimize storage and lookup speed.
In preparation for a future patch to add support for INET6,
wrap the fi_check_info call with a provider specific wrapper.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>