Change size_option to a flag
Convert the current size_option into a set of flags.  Rather
than having tests run with a particular size when the size_option
is greater than the value stored in the size array, determine
if the size should be used based on whether a specific flag
has been set.  This is more straightforward to use as a developer
and more flexible.

The size 'all' option will now attempt to use all values in the
size array.  Add a DEFAULT option flag to indicate which sizes
in the table should be used in default size iteration tests.  The
default size values are updated to new values.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>