benchmarks: do inject operation as part of pingpong
Allow an inject size to be specified on input and have the
pingpong tests use inject if message qualifies for it.

Remove the inject_pingpong test, since it is now integrated
as a part of the other benchmarks.

Set inject value in hints structure, and use the value then
provided by the output fi_info structure.

Signed-off-by: Sayantan Sur <sayantan.sur@intel.com>