fabric: Define mechanism to report unknown source address data
When a domain has been configured with FI_SOURCE, completions
should return the fi_addr_t associated with the source address
for a received message.  However, it is possible that no fi_addr_t
value is available.  For example, the source address has not been
inserted into the corresponding AV.

There is currently no way for a provider to give the application
the raw source address data, so that it may be inserted into the
AV.  Define a way to report this data.

The method used is to inject a completion error into the CQ.  The
error code is set to -FI_EADDRNOTAVAIL.  And the error event data
will contain the source address.  The source address must be
reported using the same format specified in the fi_info::
addr_format field when the domain was opened.

This feature is automatically enabled when the domain has been
configured with FI_SOURCE and the application selects API
version 1.3 or higher.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>