fabric: Bump ABI to version 1.1
Prepare for changes to the core data structures, like fi_info,
by bumping the ABI from 1.0 to 1.1.  Symbol versioning will be
used to provide backwards compatibility to existing apps.

This is a staging patch.  Additional updates will be needed as
the 1.0 data structures are actually modified, in order to
provide the proper translation between structures.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>