Getinfo doesn't work for the sockets provider, if the hostname of the node doesn't resolve to IP address. Sock_getinfo uses gethostname to find the source address to use, but sock_ep_getinfo gets an error when it calls getaddrinfo. The end result is that the provider fails even though there is an IP network defined.
This patch modifies the behavior of sock_getinfo so it checks in
advance if the result of gethostname can be resolved by getaddrinfo.
If not, it uses getifaddrs to find a non-loopback network interface
which interface can be used in place of the hostname. If that fails,
it uses the loopback address (127.0.0.1).