man/fd: Document that fd's must be checked for POLLIN to be active
See issue #1674.  Current text reads:

File descriptor wait objects must be usable in select, poll, and epoll
routines.  However, a provider may signal an FD wait object by marking
it as readable, writable, or with an error.

It may actually be impossible for a provider not to have an fd always
marked as writable.  For example, many providers use pipes or socket
pairs as a signaling object for EQs.  Restrict the implementation to
only marking the socket as readable or error.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>