man/sread: Restrict use of sread calls where waitsets aren't used
Mixing fi_cq_sread and fi_eq_sread with fi_wait can result in
undefined behavior, particularly if wait objects are shared
among multiple fid's.  To avoid this, document that the sread
calls cannot be used if the CQ/EQ is configured to use a wait
set as its wait object.  Instead the fi_wait call must be
used.

This fixes issue #1533.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>