configury: use configure-based libfabric.map
In the course of implementing FI_DIRECT in the gni provider,
it was found that the current hardwired version-script isn't
really suitable.

A method is needed to generate the version-script depending
on how libfabric is configured.  There is no need to change
the base script if --enable-direct=<provider> is not specified
on the configure line.  However, if it is, there needs to
be a mechanism for a provider to add its provider specific
entry points to the global section of the version script.

This PR introduces such a capability.

With this PR, The libfabric.map is now generated using
the C preprocessor.

Only the libfabric.map.cpp file is included in the dist
tarball.  A make clean will remove the libfabric.map file.

Signed-off-by: Evan Harvey <eharvey@lanl.gov>