man/ep: Document that FI_RECV includes FI_REMOTE_WRITE
This is related to issues #1637 and #1636.  When binding a CQ
to an EP with FI_RECV, this also includes events generated
from using a remote write.

Update that FI_TRANSMIT includes FI_READ + FI_WRITE, and
FI_RECV includes FI_REMOTE_WRITE, when binding an EP to a CQ.

That is, CQ bindings are different than counter bindings.  Just
make this more explicit.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>