crontab: update to latest copy
Save a copy of ofiwg-bot's crontab from flatbed.openfabrics.org.

Disable the v1.1.x builds.  No further work is occurring on that
branch, and the fabtests master no longer builds with it (because
fabtests master uses new APIs that were recently added to libfabric
master that do not exist in libfabric 1.1.x).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>