util/cirque: Allocate buffer with cirque structure
Cirque allocates the array of user structures separate from
the cirque structure.  This is unnecessary, since an allocation
is needed anyway.  Combine the two as part of a single allocation.

This will allow users to allocate the cirque memory themselves
as part of another allocation, then format the memory for use
as a circular queue.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>