usnic: Add is_soft flag to internal CQ structure.
The sread function will differ depending on the type of wait object
associated with the CQ. This means that instead of just setting the
pointer to a static struct, the code will have to keep a copy around so
that it can modify the sread function on a per CQ basis. This also
avoids having to compare the CQ ops whenever determining whether the CQ
is soft.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>