usnic: fix compiler warning
Compiler warned that "ret" was set but not used.  Now support the
run-time error in non-debug builds (i.e., where the assert() will not
be tripped).  Also move the existing assert() beyond the debug
statement (so that something useful is printed out on the screen), and
add a symetric assert() in the poll_close() case.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>