prov/psm2: use the eq/wait/poll objects from the utility provider
Replace the local implementations of eq, waitset, and pollset with
the corresponding ones from the utility provider. The benefit is
reduction of code duplication and better support for the wait/poll
functions. The "trywait" function is also automatically supported.

Some changes are also made to the fabric and domain object due to
the dependency among various objects in the utility provider. But
overall these changes are minor.

Signed-off-by: Jianxin Xiong <jianxin.xiong@intel.com>