usnic: Make sure progress thread is awake before waiting.
If the progress thread decides that it doesn't have any work to do, then
it could start waiting on the control FD. In that scenario, functions
like fi_cq_sread and fi_wait could lead to a deadlock.

Add a new atomic variable, with semantics similar to a counting
semaphore, to the track the number of threads currently waiting. Before
waiting issue a wakeup write to the progress thread.

In the future this could be resolved in a way that performs better for
hard CQs on devices with interrupt enabled completions.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>