usnic: Use EFD_SEMAPHORE for EQ and fabric control FDs.
Commit 8c533eb removed the EFD_SEMAPHORE flag from all control
descriptors. This is alright for the CQ operations since the provider
has to handle the FD used for the hardware CQs that does not support
counting. This means that the FD is cleared upon every invocation of
trywait. For the EQs, the FD should not be cleared. The EQ operations
rely on the counting behavior of an FD opened with EFD_SEMAPHORE. This
commit adds back the EFD_SEMAPHORE to the fabric and EQ control FDs.
This fixes #1896.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>