usnic: Keep progress thread alive for CQs.
A common usage pattern for external FD use is to acquire the FD, call
trywait, and call fi_cq_read. The progress threads needs to stay alive
during the duration of the wait in order to progress receives. Keep
track of CQs that are waiting and increment the fabric wait var to keep
the thread alive. Decrement once again upon the first fi_cq_read for
each CQ.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>