configure.ac: use "ustar" tar format
Some of the filenames get very long in Ubuntu builds, especially when
adding vendor-specific version numbers -- so long, in fact, that the
default format for tar will not include some files in the resulting
tarball.  So instead, use the "ustar" format in tar, that allows for
longer filenames.

This Automake option (to use the "ustar" format) has been in use by
the Open MPI project (literally) for years; it is quite safe and
portable.

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>