index: Replace embedded docs with links and re-org web page
This restructures the libfabric web page.  A link to the latest
release is moved to the top of the page.  The documentation portion
is updated to remove embedded documents, with links provided instead.

The general content is left intact, with things mostly just
restructured to clean up the page and shrink it to something more
easily digested.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>