usnic: Fix handling of connection request data.
This fixes a segmentation fault when handling CM connections that
contain data (param argument to fi_connect/fi_accept). The buffer
pointer was not properly being updated and a later conditional is
dependent upon it. Since it was not being updated, the size of the data
was not being converted from network to host long. This led to a seg
fault after giving memcpy an impracticly large size.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>