man/fi_endpoint: Fix wording for counting RMA events
Counters track completions, not submissions.  The current wording
on the man page is confusing as it refers to the initiation of
the operation, rather than the completion.  The initiation is simply
trying to convey which side the operation was posted on, and is not
related to when the counter is updated.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>