usnic: Fix invalid completion lengths.
The FI_EP_RDM and FI_EP_MSG implementations in the usNIC provider use
fields in the work requests (rd_length and ms_length) to keep track of
how many bytes have been received. Some of the receive functions were
incorrectly initializing these lengths to the full length of the
receive. This resulted in a completion length that counted the length
twice. Initialize them to 0 instead.

Fixes #1832.

Signed-off-by: Ben Turrubiates <bturrubi@cisco.com>