headers: do not install provider-only header files
Do not install fi_log.h and fi_prov.h.  These files are intended for
providers only, and do not need to be installed where applications can
see them.

These files do include interfaces with ABI guarantees, however, so
keep them under include/rdma (because the files in include/ are not
publicly exported from libfabric, and therefore are not subject to ABI
guarantees).

Signed-off-by: Jeff Squyres <jsquyres@cisco.com>