Verbs: Add SRQ support through SRX context
This patch adds the support of Shared-Receive Queues to the
Verbs provider.

The SRQ is allocated using fi_rx_context(), which creates a new 'fid_ep'
corresponding to the SRQ. This SRQ endpoint can then be associated
to an endpoint using fi_ep_bind().

Once an endpoint is associated to a SRX context, the user can
post receive buffers to the SRX using 'fi_recv'.

Signed-off-by: Sean Hefty sean.hefty@intel.com