Verbs: Add SRX support to FI_MSG endpoint
This patch enables the support of SRX context (through
the Shared-Receive Queues) to the FI_MSG endpoints.

SRX allows to share a single receive queue among multiple
connections. This aims at reducing memory consumption at
scale.

When a SRQ is associated to a QP, the number of outstanding WRs
in the QP's receive queue is automatically set to 0.

Once a SRX context is associated to an endpoint, it is then
forbidden to post receive buffers to that endpoint (ENOSYS is
returned in this case).
Only posting receive buffers to the endpoint corresponding to
the SRX is allowed.

Signed-off-by: Sean Hefty sean.hefty@intel.com