man/atomic: Document that long double is not documented :)
The size of a long double is implementation specific, including
how many bytes it consumes and how many bits are used.  Until
we come up with a way for an app to determine the sizes,
document that the size is platform/compiler/provider/hardware
specific and wish the developers good luck using it.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>