docs: Add conception of an OFI programmer's guide
Add a new document that will eventually become a book
that describes libfabric in detail.

Provide a rough outline for the material.  The intent is to
pull content from other documents and the man pages, but add
considerable details regarding why the API is defined the way
that it is.  It will assume that the reader is familiar with
sockets programming, and make references to sockets where
possible in order to aid understanding of the API.

The book will be written using markdown and maintained as
part of the repository.  It can be published as html on
github and also published from gitbooks.

Signed-off-by: Sean Hefty <sean.hefty@intel.com>